static methods in ColorUtil
